# docs/static/openapi.yaml
openapi: 3.0.1
info:
  title: Szampchat server api
  version: "0.1"
servers:
- url: http://localhost:8081/api
  description: Local server
paths:
...
  /users:
    post:
      tags:
      - User
      summary: Register user
      description: |
        Used to create user object in backend server on first login.
        Only route that doesn't throw 419 error.
        To obtain bearer token, check Basics/Authentication page or KeyCloak documentation.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
    delete:
      tags:
      - User
      summary: Delete user
      operationId: deleteUser
      responses:
        "204":
          description: No Content
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
...
  /communities/{communityId}/roles:
    post:
      tags:
      - Role
      summary: Create role TODO
      operationId: createRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreateDTO"
        required: true
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "419":
          description: User not created
        "403":
          description: User is not an owner of community or resource is not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
...
  /users/description:
    patch:
      tags:
      - User
      summary: Edit description?
      operationId: editDescription
      parameters:
      - name: descriptionDTO
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/UserDescriptionDTO"
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
  /users/avatar:
    patch:
      tags:
      - User
      summary: Edit avatar? TODO
      operationId: editAvatar
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
      security:
      - OAuthSecurity: []
...
  /communities/{communityId}/roles/{roleId}:
    patch:
      tags:
      - Role
      summary: Edit role TODO
      operationId: editRole
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreateDTO"
        required: true
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "419":
          description: User not created
        "403":
          description: User is not a member of channel's community or resource is
            not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
...
  /users/{userId}:
    get:
      tags:
      - User
      summary: Get user
      operationId: getUser
      parameters:
      - name: userId
        in: path
        description: ID of user
        required: true
        schema:
          type: integer
          format: int64
        example: "4501678724218880"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
  /users/me:
    get:
      tags:
      - User
      summary: Get current user
      description: |
        Retrieves current user's data
      operationId: getMe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "419":
          description: User not created
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
  /roles/{roleId}:
    get:
      tags:
      - Role
      summary: Get role
      operationId: getRole
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "419":
          description: User not created
        "403":
          description: User is not a member of role's community or resource is not
            available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
    delete:
      tags:
      - Role
      summary: Delete role TODO
      operationId: deleteRole
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "419":
          description: User not created
        "403":
          description: User is not an owner of community or resource is not available
        "401":
          description: Incorrect token scope (token not provided?)
      security:
      - OAuthSecurity: []
...
components:
  schemas:
...
    UserCreateDTO:
      type: object
      properties:
        username:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID of user
          format: int64
          example: 4501678724218880
        name:
          type: string
          description: User name
          example: User123
        imageUrl:
          type: string
          description: Relative path to image (or uuid)
          example: img/userimage.jpg
        description:
          type: string
          description: Short message which describes user
          example: My description
...
    RoleCreateDTO:
      type: object
    PermissionOverwrites:
      type: object
      properties:
        permissionOverwriteData:
          type: integer
          format: int64
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        permission:
          $ref: "#/components/schemas/PermissionOverwrites"
        community:
          type: integer
          format: int64
...
    UserDescriptionDTO:
      type: object
      properties:
        description:
          type: string
...
  securitySchemes:
    OAuthSecurity:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://url.com/auth
          scopes: {}